back end script
Added the code for the back end php...trying to look through it to make sense of it but wanted to get it to you guys to see if you could get anything out of it.

I have the log in script on the html page just in case this proves to be hard to implement.